+++ crates/wdl-ast/src/experimental/v1/decls.rs
@@ -2,12 +2,11 @@
 
 use std::fmt;
 
-use rowan::ast::support;
-use rowan::ast::support::child;
-use rowan::ast::AstNode;
-
 use super::Expr;
+use crate::experimental::support;
+use crate::experimental::support::child;
 use crate::experimental::token;
+use crate::experimental::AstNode;
 use crate::experimental::AstToken;
 use crate::experimental::Ident;
 use crate::experimental::SyntaxKind;