+++ crates/wdl-ast/src/experimental/v1/struct.rs
@@ -1,11 +1,10 @@
 //! V1 AST representation for struct definitions.
 
-use rowan::ast::support::children;
-use rowan::ast::AstChildren;
-use rowan::ast::AstNode;
-
 use super::UnboundDecl;
+use crate::experimental::support::children;
 use crate::experimental::token;
+use crate::experimental::AstChildren;
+use crate::experimental::AstNode;
 use crate::experimental::Ident;
 use crate::experimental::SyntaxKind;
 use crate::experimental::SyntaxNode;