+++ crates/wdl-ast/src/experimental/v1/validation/numbers.rs
@@ -1,11 +1,10 @@
 //! Validation of number literals in a V1 AST.
 
-use rowan::ast::support;
-use rowan::ast::AstNode;
-
+use crate::experimental::support;
 use crate::experimental::v1::Expr;
 use crate::experimental::v1::LiteralExpr;
 use crate::experimental::v1::Visitor;
+use crate::experimental::AstNode;
 use crate::experimental::AstToken;
 use crate::experimental::Diagnostic;
 use crate::experimental::Diagnostics;